Removed build for demo & moved items to assets
